Rename ICES agent files
